requirements.txt
# AI Agent Framework
langchain-core>=0.3.0
langgraph>=0.2.0

# Gmail & Calendar API
google-api-python-client>=2.100.0
google-auth>=2.0.0
google-auth-oauthlib>=1.2.0

# UI
streamlit>=1.40.0

# Utilities
python-dotenv>=1.0.0
requests>=2.28.0

# Local inference
llama-cpp-python>=0.1.60
